+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 2.0.1.pre
-- TODO
+- New `:max_age_fallback` (in seconds) option
 
 ## 2.0.0 / 2014-12-07
 - Rack Zippy now works cleanly with all Rack apps, including Rails apps ([#23](https://github.com/eliotsykes/rack-zippy/issues/23))